Closes #22477
Add a check so if the dtype is str is will create
an empty array type object and then pass the values.

Add test for an empty series. To chech that it fills the series
with NaN and not with 'n'.

Also add a test for cases that no string values are given.